Extract logic in sperate file
